Add comments, split source into multiple files
