+++ pre/encoding.c
@@ -37,546 +37,9 @@
 #include "pre-afgh-relic.h"
 #include <limits.h>
 
-// Finds the mod inverse of a modulo m
-int mod_inverse(bn_t res, bn_t a, bn_t m) {
-  bn_t tempGcd, temp;
-  int result = STS_OK;
-
-  bn_null(tempGcd);
-  bn_null(temp);
-
-  TRY {
-
-    bn_new(tempGcd);
-    bn_new(temp);
-
-    bn_gcd_ext(tempGcd, res, temp, a, m);
-    if (bn_sign(res) == BN_NEG) {
-      bn_add(res, res, m);
-    }
-  }
-  CATCH_ANY { result = STS_ERR; }
-  FINALLY {
-    bn_free(tempGcd);
-    bn_free(temp);
-  }
-
-  return result;
-}
-
-int pre_init() {
-  if (core_init() != STS_OK) {
-    core_clean();
-    return STS_ERR;
-  }
-
-  if (pc_param_set_any() != STS_OK) {
-    THROW(ERR_NO_CURVE);
-    core_clean();
-    return STS_ERR;
-  }
-  pc_param_print();
-  return STS_OK;
-}
-
-int pre_deinit() {
-  core_clean();
-  return STS_OK;
-}
-
-int pre_params_clear(pre_params_t params) {
-  int result = STS_OK;
-
-  TRY {
-    assert(params);
-
-    gt_free(params->Z);
-    g1_free(params->g1);
-    g2_free(params->g2);
-  }
-  CATCH_ANY { result = STS_ERR; }
-
-  return result;
-}
-
-int pre_sk_clear(pre_sk_t sk) {
-  int result = STS_OK;
-
-  TRY {
-    assert(sk);
-
-    bn_free(sk->sk);
-    bn_free(sk->inverse);
-  }
-  CATCH_ANY { result = STS_ERR; }
-
-  return result;
-}
-
-int pre_pk_clear(pre_pk_t pk) {
-  int result = STS_OK;
-
-  TRY {
-    assert(pk);
-
-    g1_free(pk->pk1);
-    g2_free(pk->pk2);
-  }
-  CATCH_ANY { result = STS_ERR; }
-
-  return result;
-}
-
-int pre_rand_message(gt_t msg) {
-  int result = STS_OK;
-
-  TRY {
-    gt_new(msg);
-    gt_rand(msg);
-  }
-  CATCH_ANY { result = STS_ERR; }
-
-  return result;
-}
-
-int pre_map_to_key(uint8_t *key, int key_len, gt_t msg) {
-  int result = STS_OK;
-
-  uint8_t *buffer;
-  size_t buff_size;
-
-  TRY {
-    buff_size = (size_t)gt_size_bin(msg, 1);
-    buffer = (uint8_t *)malloc(buff_size);
-    if (!buffer) {
-      return STS_ERR;
-    }
-    gt_write_bin(buffer, (int)buff_size, msg, 1);
-    md_kdf2(key, key_len, buffer, (int)buff_size);
-  }
-  CATCH_ANY { result = STS_ERR; }
-  FINALLY {
-    if (buffer) {
-      free(buffer);
-    }
-  };
-
-  return result;
-}
-
-// cleanup and free sub-structures in 'ciphertext' (only if already initialized)
-int pre_ciphertext_clear(pre_ciphertext_t ciphertext) {
-  int result = STS_OK;
-
-  TRY {
-    assert(ciphertext);
-    assert(ciphertext->C1);
-    assert(ciphertext->C2_G1 || ciphertext->C2_GT);
-
-    if (ciphertext->group ==
-        PRE_REL_CIPHERTEXT_IN_G_GROUP) { // test to detect if the structure is
-                                         // already initialized); could it fail?
-      gt_free(ciphertext->C1);
-      g1_free(ciphertext->C2_G1);
-      ciphertext->group = '\0';
-    } else if (ciphertext->group == PRE_REL_CIPHERTEXT_IN_GT_GROUP) {
-      gt_free(ciphertext->C1);
-      gt_free(ciphertext->C2_GT);
-      ciphertext->group = '\0';
-    }
-  }
-  CATCH_ANY { result = STS_ERR; }
-
-  return result;
-}
-
-int pre_generate_params(pre_params_t params) {
-  int result = STS_OK;
-
-  g1_null(params->g1);
-  g2_null(params->g2);
-  gt_null(params->Z);
-  bn_null(params->g1_ord);
-
-  TRY {
-    g1_new(params->g1);
-    g2_new(params->g2);
-    gt_new(params->Z);
-    bn_new(params->g1_ord);
-
-    g1_get_gen(params->g1);
-    g2_get_gen(params->g2);
-
-    /* pairing Z = e(g,g)*/
-    pc_map(params->Z, params->g1, params->g2);
-
-    g1_get_ord(params->g1_ord);
-  }
-  CATCH_ANY {
-    result = STS_ERR;
-
-    g1_null(params->g1);
-    g2_null(params->g2);
-    gt_null(params->Z);
-    bn_null(params->g1_ord);
-  };
-
-  return result;
-}
-
-int pre_generate_sk(pre_params_t params, pre_sk_t sk) {
-  int result = STS_OK;
-
-  bn_null(sk->sk);
-  bn_null(sk->inverse);
-
-  TRY {
-    bn_new(sk->sk);
-    bn_new(sk->inverse);
-
-    // generate a random value, a, as secret key
-    bn_rand_mod(sk->sk, params->g1_ord);
-
-    // compute 1/a mod n for use later
-    mod_inverse(sk->inverse, sk->sk, params->g1_ord);
-  }
-  CATCH_ANY {
-    result = STS_ERR;
-
-    bn_null(sk->sk);
-    bn_null(sk->inverse);
-  };
-
-  return result;
-}
-
-int pre_generate_pk(pre_params_t params, pre_sk_t sk, pre_pk_t pk) {
-  int result = STS_OK;
-
-  g1_null(pk->pk1);
-  g2_null(pk->pk2);
-
-  TRY {
-    g1_new(pk->pk1);
-    g2_new(pk->pk2);
-
-    // compute the public key as pk1 = g1^a, pk2 = g2^a
-    g1_mul_gen(pk->pk1, sk->sk);
-    g2_mul_gen(pk->pk2, sk->sk);
-  }
-  CATCH_ANY {
-    result = STS_ERR;
-
-    g1_null(pk->pk1);
-    g2_null(pk->pk2);
-  };
-
-  return result;
-}
-
-int hash_pk(bn_t hash, g1_t g1_hash, g2_t g2_hash, pre_pk_t pk) { 
-  int result = STS_ERR;
-
-  int size;
-
-  bn_null(hash);
-  g1_null(g1_hash);
-  g2_null(g2_hash);
-  TRY {
-    assert(pk);
-
-    bn_new(hash);
-    g1_new(g1_hash);
-    g2_new(g2_hash);
-
-    size = g1_size_bin(pk->pk1, 1);
-    uint8_t buf[size], hash_vector[64];
-    g1_write_bin(buf, size, pk->pk1, 1);
-    md_map_sh512(hash_vector, buf, size);
-    bn_read_bin(hash, hash_vector, 64);
-
-    g1_mul_gen(g1_hash, hash);
-    g2_mul_gen(g2_hash, hash);
-  }
-  CATCH_ANY {
-    result = STS_ERR;
-
-    bn_null(hash_int);
-    g1_null(g1_hash);
-    g2_null(g2_hash);
-  }
-  FINALLY {
-    bn_free(hash_int);
-  }
-
-  return result;
-}
-
-int pre_derive_next_pk(pre_pk_t pk) {
-  int result = STS_OK;
-
-  bn_t hash;
-  g1_t g1_hash;
-  g2_t g2_hash;
-
-  TRY {
-    assert(pk);
-
-    hash_pk(hash, g1_hash, g2_hash, pk);
-
-    g1_add(pk->pk1, pk->pk1, g1_hash);
-    g2_add(pk->pk2, pk->pk2, g2_hash);
-  }
-  CATCH_ANY {
-    result = STS_ERR;
-
-    bn_free(hash);
-    g1_free(g1_hash);
-    g2_free(g2_hash);
-  }
-  FINALLY {
-    bn_free(hash);
-    g1_free(g1_hash);
-    g2_free(g2_hash);
-  }
-
-  return result;
-}
-
-int pre_derive_next_keypair(pre_sk_t sk, pre_pk_t pk) {
-  int result = STS_OK;
-
-  bn_t hash;
-  g1_t g1_hash;
-  g2_t g2_hash;
-
-  TRY {
-    assert(pk);
-
-    hash_pk(hash, g1_hash, g2_hash, pk);
-
-    g1_add(pk->pk1, pk->pk1, g1_hash);
-    g2_add(pk->pk2, pk->pk2, g2_hash);
-
-    bn_add(sk->sk, sk->sk, hash);
-  }
-  CATCH_ANY {
-    result = STS_ERR;
-
-    bn_free(hash);
-    g1_free(g1_hash);
-    g2_free(g2_hash);
-  }
-  FINALLY {
-    bn_free(hash);
-    g1_free(g1_hash);
-    g2_free(g2_hash);
-  }
-
-  return result;
-}
-
-int pre_generate_token(pre_token_t token, pre_params_t params, pre_sk_t sk, pre_pk_t pk) {
-  int result = STS_OK;
-
-  g2_null(token->token);
-  TRY {
-    assert(token);
-    assert(params);
-    assert(sk);
-    assert(pk);
-
-    g2_new(token->token);
-
-    /* g^b ^ 1/a*/
-    g2_mul(token->token, pk->pk2, sk->inverse);
-  }
-  CATCH_ANY {
-    result = STS_ERR;
-    g2_null(token->token);
-  };
-
-  return result;
-}
-
-int pre_encrypt(pre_ciphertext_t ciphertext, pre_params_t params, pre_pk_t pk, gt_t plaintext) {
-  int result = STS_OK;
-
-  bn_t r;
-
-  bn_null(r);
-
-  TRY {
-    bn_new(r);
-
-    gt_new(ciphertext->C1);
-    g1_new(ciphertext->C2_G1);
-
-    assert(ciphertext);
-    assert(params);
-    assert(pk);
-
-    ciphertext->group = PRE_REL_CIPHERTEXT_IN_G_GROUP;
-
-    /*
-     * First Level Encryption
-     * c = (c1, c2)     c1, c2 \in G
-     *      c1 = Z^ar = e(g,g)^ar = e(g^a,g^r) = e(pk_a, g^r)
-     *      c2 = m·Z^r
-     */
-    /* Compute C1 part: MZ^r*/
-    /* random r in Zn  (re-use r) */
-    bn_rand_mod(r, params->g1_ord);
-    while (bn_is_zero(r)) {
-      bn_rand_mod(r, params->g1_ord);
-    }
-
-    /* Z^r */
-    gt_exp(ciphertext->C1, params->Z, r);
-
-    /* Z^r*m */
-    gt_mul(ciphertext->C1, ciphertext->C1, plaintext);
-
-    /* Compute C2 part: G^ar = pk ^r*/
-    /* g^ar = pk^r */
-    g1_mul(ciphertext->C2_G1, pk->pk1, r);
-  }
-  CATCH_ANY {
-    result = STS_ERR;
-    gt_null(ciphertext->C1);
-    g1_null(ciphertext->C2_G1);
-  }
-  FINALLY {
-    bn_free(r);
-  }
-
-  return result;
-}
-
-int pre_decrypt(gt_t plaintext, pre_params_t params, pre_sk_t sk, pre_ciphertext_t ciphertext) {
-  int result = STS_OK;
-
-  g2_t t1;
-  gt_t t0;
-
-  gt_null(t0);
-  g2_null(t1);
-  gt_null(plaintext);
-
-  TRY {
-    gt_new(t0);
-    g2_new(t1);
-    gt_new(plaintext);
-
-    assert(params);
-    assert(sk);
-    assert(ciphertext);
-    assert((ciphertext->group == PRE_REL_CIPHERTEXT_IN_G_GROUP) ||
-           (ciphertext->group == PRE_REL_CIPHERTEXT_IN_GT_GROUP));
-
-    /*
-     * M = (M.Z^r) / e(G^ar, G^1/a)
-     * M = C1 / e(C2, G^1/a)
-     */
-
-    if (ciphertext->group == PRE_REL_CIPHERTEXT_IN_G_GROUP) {
-      /* g ^ 1/a*/
-      g2_mul(t1, params->g2, sk->inverse);
-
-      /* e(g^ar, g^-a) = Z^r */
-      pc_map(plaintext, ciphertext->C2_G1, t1);
-      // pbc_pmesg(3, "Z^r: %B\n", t2);
-    } else {
-      /* C2 = Z^ar
-       * Compute: Z^ar^(1/a)*/
-      if (bn_is_zero(sk->inverse)) {
-        gt_set_unity(plaintext);
-      } else {
-        gt_exp(plaintext, ciphertext->C2_GT, sk->inverse);
-      }
-    }
-
-    /* C1 / e(C2, g^a^-1) or C1/C2^(1/a) */
-    gt_inv(t0, plaintext);
-    gt_mul(plaintext, ciphertext->C1, t0);
-  }
-  CATCH_ANY { result = STS_ERR; }
-  FINALLY {
-    gt_free(t0);
-    g1_free(t1);
-  }
-
-  return result;
-}
-
-int pre_apply_token(pre_token_t token, pre_ciphertext_t res,
-                 pre_ciphertext_t ciphertext) {
-  int result = STS_OK;
-
-  TRY {
-    assert(token);
-    assert(res);
-    assert(ciphertext);
-    assert(ciphertext->group == PRE_REL_CIPHERTEXT_IN_G_GROUP);
-    pre_ciphertext_clear(res);
-
-    gt_new(res->C1);
-    gt_new(res->C2_GT);
-
-    gt_copy(res->C1, ciphertext->C1);
-
-    /* level2: C2 = g^ar */
-    /* level1: C2 = e(g^ar, g^(b/a) = Z^br*/
-    pc_map(res->C2_GT, ciphertext->C2_G1, token->token);
-
-    res->group = PRE_REL_CIPHERTEXT_IN_GT_GROUP;
-  }
-  CATCH_ANY {
-    result = STS_ERR;
-    gt_null(res->C1);
-    gt_null(res->C2_GT);
-  }
-
-  return result;
-}
-
-int get_encoded_msg_size(gt_t msg) {
-  return gt_size_bin(msg, 1);
-}
-
-int encode_msg(char *buff, int size, gt_t msg) {
-  int next_size = gt_size_bin(msg, 1);
-  if (size < next_size) {
-    return STS_ERR;
-  }
-  gt_write_bin((uint8_t *)buff, next_size, msg, 1);
-  return STS_OK;
-}
-
-int decode_msg(gt_t msg, char *buff, int size) {
-  gt_new(msg);
-  gt_read_bin(msg, (uint8_t *)buff, size);
-  return STS_OK;
-}
-
-int get_encoded_token_size(pre_token_t token) {
-  return g2_size_bin(token->token, 1);
-}
-
-int encode_token(char *buff, int size, pre_token_t token) {
-  int next_size = g2_size_bin(token->token, 1);
-  if (size < next_size) {
-    return STS_ERR;
-  }
-  g2_write_bin((uint8_t *)buff, next_size, token->token, 1);
-  return STS_OK;
-}
-
-int decode_token(pre_token_t token, char *buff, int size) {
-  g2_new(token->token);
-  g2_read_bin(token->token, (uint8_t *)buff, size);
-  return STS_OK;
-}
+////////////////////////////////////////
+//    Encoding/Decoding Functions     //
+////////////////////////////////////////
 
 void write_size(char *buffer, int size) {
   buffer[0] = (char)((size >> 8) & 0xFF);
@@ -843,6 +306,44 @@ int decode_pk(pre_pk_t pk, char *buff, int size) {
   return STS_OK;
 }
 
+int get_encoded_token_size(pre_token_t token) {
+  return g2_size_bin(token->token, 1);
+}
+
+int encode_token(char *buff, int size, pre_token_t token) {
+  int next_size = g2_size_bin(token->token, 1);
+  if (size < next_size) {
+    return STS_ERR;
+  }
+  g2_write_bin((uint8_t *)buff, next_size, token->token, 1);
+  return STS_OK;
+}
+
+int decode_token(pre_token_t token, char *buff, int size) {
+  g2_new(token->token);
+  g2_read_bin(token->token, (uint8_t *)buff, size);
+  return STS_OK;
+}
+
+int get_encoded_msg_size(gt_t msg) {
+  return gt_size_bin(msg, 1);
+}
+
+int encode_msg(char *buff, int size, gt_t msg) {
+  int next_size = gt_size_bin(msg, 1);
+  if (size < next_size) {
+    return STS_ERR;
+  }
+  gt_write_bin((uint8_t *)buff, next_size, msg, 1);
+  return STS_OK;
+}
+
+int decode_msg(gt_t msg, char *buff, int size) {
+  gt_new(msg);
+  gt_read_bin(msg, (uint8_t *)buff, size);
+  return STS_OK;
+}
+
 int get_encoded_cipher_size(pre_ciphertext_t cipher) {
   int size = 1;
   size += gt_size_bin(cipher->C1, 1) + ENCODING_SIZE;
@@ -938,20 +439,5 @@ int decode_cipher(pre_ciphertext_t cipher, char *buff, int size) {
     gt_read_bin(cipher->C2_GT, (uint8_t *)curr, next_size);
   }
 
-  return STS_OK;
-}
-
-int pre_cipher_clear(pre_ciphertext_t cipher) {
-  gt_free(cipher->C1);
-  if (cipher->group == PRE_REL_CIPHERTEXT_IN_G_GROUP) {
-    g1_free(cipher->C2_G1);
-  } else {
-    gt_free(cipher->C2_GT);
-  }
-  return STS_OK;
-}
-
-int pre_token_clear(pre_token_t token) {
-  g2_free(token->token);
   return STS_OK;
 }
